feat(get_settings_profile): create default matrix of profile values
Users could run the code and see how the structure could be changed to
add profile parameters. A matrix is preferred to a data frame so that you
know that the columns will all be numeric. Then the rownames are
the real parameter names. Matrices are quicker to access and take up less
space than data frames. Feel free to remove this function, I just thought
it was the best way to share an example. The profile_wrapper can act on
the rows.