fix(retro_wrapper): round mohn's rho values in table
Use round() but had to unlist() the listed column otherwise it
didn't think that it was numeric. Kind of weird. But it works.
Move spread to after manipulating the data frame.

No longer use longtable = TRUE b/c it is not needed for accessability and
it is best if this table is not split across pages. But, right align
all the values because there can be negative numbers which makes left or
center align annoying.